Edit Repair class and move it into record-folder. Change Id type to Guid, edit corresponding classes. Fix namespaces.
