+++ CarRental.BussinessLayer/Managers/RepairManager.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using CarRental.Data.Models;
+using CarRental.Data.Models.RecordTypes;
 using CarRental.Data.Models.Automobile;
 using CarRental.BussinessLayer.Validators;
 
@@ -33,7 +34,7 @@ namespace CarRental.BussinessLayer.Managers
 
             Repair repair = new Repair
             {
-                 Id = car.Brand.ToUpper() + car.Model.ToUpper() + DateTime.Now.ToString().ToUpper() + mechanic.Id.ToString().Substring(33).ToUpper(),
+                 Id = Guid.NewGuid(),
                  Date = DateTime.Now,
                  CarId = car.CarId,
                  CarBrand = car.Brand,
@@ -74,7 +75,7 @@ namespace CarRental.BussinessLayer.Managers
         {
             NullValidator.CheckNull(repairs);
 
-            Repair repair = repairs.Find(x => x.Id.Contains(id));
+            Repair repair = repairs.Find(x => (x.Id.CompareTo(id) == 0));
 
             NullValidator.CheckNull(repair);
 
@@ -102,7 +103,7 @@ namespace CarRental.BussinessLayer.Managers
         {
             NullValidator.CheckNull(repairs);
 
-            repairs.RemoveAt(repairs.IndexOf(repairs.Find(x => x.Id.Contains(id))));
+            repairs.RemoveAt(repairs.IndexOf(repairs.Find(x => x.Id.CompareTo(id) == 0)));
         }
     }
 }
