+++ CarRental.PresentationLayer/Models/Mechanic.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using CarRental.Data.Models.RecordTypes;
 
 namespace CarRental.Data.Models
 {
